// src/main/resources/site/x-data/layerMigration/layerMigration.ts
// WARNING: This file was automatically generated by "no.item.xp.codegen". You may lose your changes if you edit it.
export interface LayerMigration {
  /**
   * Type (live eller arkivert)
   */
  targetReferenceType?: "live" | "archived";

  /**
   * Content id (viser til arkivert innhold fra live innhold og vice versa)
   */
  contentId: string;

  /**
   * Repo id (viser til arkivert innhold fra live innhold og vice versa)
   */
  repoId: string;

  /**
   * Locale (viser til arkivert innhold fra live innhold og vice versa)
   */
  locale: string;

  /**
   * Timestamp for migrering
   */
  ts: string;
}